+++ src/lib_store/unix/test/alpha_utils.ml
@@ -262,7 +262,8 @@ module Forge = struct
 
   let make_contents ~payload_hash ~payload_round
       ?(proof_of_work_nonce = default_proof_of_work_nonce)
-      ?(liquidity_baking_toggle_vote = Liquidity_baking.LB_pass)
+      ?(liquidity_baking_toggle_vote = Toggle_votes.Toggle_vote_pass)
+      ?(adaptive_inflation_vote = Toggle_votes.Toggle_vote_pass)
       ~seed_nonce_hash () =
     Block_header.
       {
@@ -270,7 +271,11 @@ module Forge = struct
         payload_round;
         proof_of_work_nonce;
         seed_nonce_hash;
-        liquidity_baking_toggle_vote;
+        toggle_votes =
+          {
+            liquidity_baking_vote = liquidity_baking_toggle_vote;
+            adaptive_inflation_vote;
+          };
       }
 
   let make_shell ~level ~predecessor ~timestamp ~fitness ~operations_hash