+++ src/proto_alpha/lib_delegate/client_daemon.ml
@@ -72,7 +72,7 @@ let await_protocol_start (cctxt : #Protocol_client_context.full) ~chain =
 
 module Baker = struct
   let run (cctxt : Protocol_client_context.full) ?minimal_fees
-      ?minimal_nanotez_per_gas_unit ?minimal_nanotez_per_byte ?liquidity_baking
+      ?minimal_nanotez_per_gas_unit ?minimal_nanotez_per_byte ?votes
       ?extra_operations ?dal_node_endpoint ?force_apply ?context_path ~chain
       ~keep_alive delegates =
     let process () =
@@ -83,7 +83,7 @@ module Baker = struct
           ?minimal_fees
           ?minimal_nanotez_per_gas_unit
           ?minimal_nanotez_per_byte
-          ?liquidity_baking
+          ?votes
           ?extra_operations
           ?dal_node_endpoint
           ?force_apply