+++ src/proto_alpha/lib_delegate/client_daemon.mli
@@ -32,7 +32,7 @@ module Baker : sig
     ?minimal_fees:Protocol.Alpha_context.Tez.t ->
     ?minimal_nanotez_per_gas_unit:Q.t ->
     ?minimal_nanotez_per_byte:Q.t ->
-    ?liquidity_baking:Baking_configuration.liquidity_baking_config ->
+    ?votes:Baking_configuration.toggle_votes_config ->
     ?extra_operations:Baking_configuration.Operations_source.t ->
     ?dal_node_endpoint:Uri.t ->
     ?force_apply:bool ->