+++ src/proto_alpha/lib_protocol/block_header_repr.mli
@@ -30,8 +30,7 @@ type contents = {
   payload_round : Round_repr.t;
   seed_nonce_hash : Nonce_hash.t option;
   proof_of_work_nonce : bytes;
-  liquidity_baking_toggle_vote :
-    Liquidity_baking_repr.liquidity_baking_toggle_vote;
+  toggle_votes : Toggle_votes_repr.toggle_votes;
 }
 
 type protocol_data = {contents : contents; signature : Signature.t}