+++ src/proto_alpha/lib_protocol/test/helpers/block.mli
@@ -72,7 +72,8 @@ module Forge : sig
   val contents :
     ?proof_of_work_threshold:Int64.t ->
     ?seed_nonce_hash:Nonce_hash.t ->
-    ?liquidity_baking_toggle_vote:Liquidity_baking.liquidity_baking_toggle_vote ->
+    ?liquidity_baking_toggle_vote:Toggle_votes.toggle_vote ->
+    ?adaptive_inflation_vote:Toggle_votes.toggle_vote ->
     payload_hash:Block_payload_hash.t ->
     payload_round:Round.t ->
     Block_header.shell_header ->
@@ -90,7 +91,8 @@ module Forge : sig
     ?policy:baker_policy ->
     ?timestamp:Timestamp.time ->
     ?operations:Operation.packed list ->
-    ?liquidity_baking_toggle_vote:Liquidity_baking.liquidity_baking_toggle_vote ->
+    ?liquidity_baking_toggle_vote:Toggle_votes.toggle_vote ->
+    ?adaptive_inflation_vote:Toggle_votes.toggle_vote ->
     t ->
     header tzresult Lwt.t
 
@@ -208,7 +210,8 @@ val bake :
   ?timestamp:Timestamp.time ->
   ?operation:Operation.packed ->
   ?operations:Operation.packed list ->
-  ?liquidity_baking_toggle_vote:Liquidity_baking.liquidity_baking_toggle_vote ->
+  ?liquidity_baking_toggle_vote:Toggle_votes.toggle_vote ->
+  ?adaptive_inflation_vote:Toggle_votes.toggle_vote ->
   ?check_size:bool ->
   t ->
   t tzresult Lwt.t
@@ -217,7 +220,8 @@ val bake :
 val bake_n :
   ?baking_mode:baking_mode ->
   ?policy:baker_policy ->
-  ?liquidity_baking_toggle_vote:Liquidity_baking.liquidity_baking_toggle_vote ->
+  ?liquidity_baking_toggle_vote:Toggle_votes.toggle_vote ->
+  ?adaptive_inflation_vote:Toggle_votes.toggle_vote ->
   int ->
   t ->
   block tzresult Lwt.t
@@ -226,7 +230,8 @@ val bake_n :
 val bake_until_level :
   ?baking_mode:baking_mode ->
   ?policy:baker_policy ->
-  ?liquidity_baking_toggle_vote:Liquidity_baking.liquidity_baking_toggle_vote ->
+  ?liquidity_baking_toggle_vote:Toggle_votes.toggle_vote ->
+  ?adaptive_inflation_vote:Toggle_votes.toggle_vote ->
   Raw_level.t ->
   t ->
   block tzresult Lwt.t
@@ -236,7 +241,8 @@ val bake_until_level :
 val bake_n_with_all_balance_updates :
   ?baking_mode:baking_mode ->
   ?policy:baker_policy ->
-  ?liquidity_baking_toggle_vote:Liquidity_baking.liquidity_baking_toggle_vote ->
+  ?liquidity_baking_toggle_vote:Toggle_votes.toggle_vote ->
+  ?adaptive_inflation_vote:Toggle_votes.toggle_vote ->
   int ->
   t ->
   (block * Alpha_context.Receipt.balance_updates) tzresult Lwt.t
@@ -259,10 +265,11 @@ val bake_n_with_origination_results :
 val bake_n_with_liquidity_baking_toggle_ema :
   ?baking_mode:baking_mode ->
   ?policy:baker_policy ->
-  ?liquidity_baking_toggle_vote:Liquidity_baking.liquidity_baking_toggle_vote ->
+  ?liquidity_baking_toggle_vote:Toggle_votes.toggle_vote ->
+  ?adaptive_inflation_vote:Toggle_votes.toggle_vote ->
   int ->
   t ->
-  (block * Alpha_context.Liquidity_baking.Toggle_EMA.t) tzresult Lwt.t
+  (block * Alpha_context.Toggle_EMA.t) tzresult Lwt.t
 
 val current_cycle : t -> Cycle.t tzresult Lwt.t