+++ src/proto_alpha/lib_protocol/toggle_EMA.mli
@@ -2,7 +2,7 @@
 (*                                                                           *)
 (* Open Source License                                                       *)
 (* Copyright (c) 2021 Tocqueville Group, Inc. <contact@tezos.com>            *)
-(* Copyright (c) 2022 Nomadic Labs <contact@nomadic-labs.com>                *)
+(* Copyright (c) 2023 Nomadic Labs <contact@nomadic-labs.com>                *)
 (*                                                                           *)
 (* Permission is hereby granted, free of charge, to any person obtaining a   *)
 (* copy of this software and associated documentation files (the "Software"),*)
@@ -24,41 +24,22 @@
 (*                                                                           *)
 (*****************************************************************************)
 
-(** Options available for the Liquidity Baking per-block vote *)
-
-type liquidity_baking_toggle_vote = LB_on | LB_off | LB_pass
-
-val liquidity_baking_toggle_vote_encoding :
-  liquidity_baking_toggle_vote Data_encoding.encoding
-
 (** Exponential moving average of toggle votes. Represented as an int32 between
-    0 and 2,000,000. It is an exponential moving average of the [LB_off] votes
-    over a window of the most recent 2000 blocks that did not vote [LB_pass]. *)
+    0 and 2,000,000,000. It is an exponential moving average of the "off" votes
+    over a window of the most recent 2000 blocks that did not vote "pass". *)
 
-module Toggle_EMA : sig
-  type t
+type t
 
-  val of_int32 : Int32.t -> t tzresult Lwt.t
+val of_int32 : Int32.t -> t tzresult Lwt.t
 
-  val zero : t
+val zero : t
 
-  val to_int32 : t -> Int32.t
+val to_int32 : t -> Int32.t
 
-  val encoding : t Data_encoding.t
+val encoding : t Data_encoding.t
 
-  val ( < ) : t -> Int32.t -> bool
-end
+val ( < ) : t -> Int32.t -> bool
 
-(** [compute_new_ema ~toggle_vote old_ema] returns the value [new_ema] of the
-    exponential moving average [old_ema] updated by the vote [toggle_vote].
+val update_ema_off : t -> t
 
-    It is updated as follows:
-    - if [toggle_vote] is [LB_pass] then [new_ema] = [old_ema],
-    - if [toggle_vote] is [LB_off], then [new_ema] = (1999 * ema[n] // 2000) + 1,000,000,
-    - if [toggle_vote] is [LB_on], then [new_ema] = (1999 * ema[n] // 2000).
-
-    The multiplication is performed in [Z.t] to avoid overflows, division is
-    rounded toward 1,000,000,000 (the middle of the interval).
-    *)
-val compute_new_ema :
-  toggle_vote:liquidity_baking_toggle_vote -> Toggle_EMA.t -> Toggle_EMA.t
+val update_ema_on : t -> t